refactored code and removed unnecessary classes
